backend/requirements.txt
fastapi
uvicorn[standard]==0.24.0
pydantic
# pyCairo
# rlPyCairo
svglib==1.0.1
reportlab==3.5.59
requests
beautifulsoup4
pandas
networkx
matplotlib
pillow
# svgwrite
svgpathtools
